refactor: Modal 내부 Button을 공통 Button 컴포넌트로 분리
