+++ src/shared/components/Modal/Button/ConfirmButton.tsx
@@ -1,12 +1,18 @@
-import { ConfirmButtonStyle } from "../styles";
-import ModalButton from "./Button";
+import { ConfirmButtonStyle } from '../styles';
+import ModalButton from '../../Button';
 
 export interface ConfirmButtonProps extends React.ComponentProps<typeof ModalButton> {
   onConfirm?: () => void;
   onHide: () => void;
 }
 
-const ConfirmButton = ({ children = "확인", onConfirm, onHide, disabled, ...rest }: ConfirmButtonProps) => {
+const ConfirmButton = ({
+  children = '확인',
+  onConfirm,
+  onHide,
+  disabled,
+  ...rest
+}: ConfirmButtonProps) => {
   const handleClick = () => {
     if (disabled) return;
     onConfirm?.(); // 사용자 정의 동작
@@ -14,7 +20,13 @@ const ConfirmButton = ({ children = "확인", onConfirm, onHide, disabled, ...re
   };
 
   return (
-    <ModalButton css={ConfirmButtonStyle(disabled)} onClick={handleClick} disabled={disabled}  aria-label="확인" {...rest}>
+    <ModalButton
+      css={ConfirmButtonStyle(disabled)}
+      onClick={handleClick}
+      disabled={disabled}
+      aria-label="확인"
+      {...rest}
+    >
       {children}
     </ModalButton>
   );
